+++ README.md
@@ -59,7 +59,7 @@
 
 - [ ] 본인만의 기준으로 일관된 코드를 작성해주세요.
 - [x] 기능 단위로 나누어 커밋을 해주세요.
-- [ ] 재사용 가능성이 높은 컴포넌트, Hooks가 보이면 재사용성을 고려하여 만들어요.
+- [x] 재사용 가능성이 높은 컴포넌트, Hooks가 보이면 재사용성을 고려하여 만들어요.
 - [x] 주문하기 페이지의 폼을 React Hook Form를 사용해서 리팩터링해요. (form 데이터를 useState로 다루는 경우는 없게 해요.)
 - [x] 받는 사람 기능 고도화
   - [x] [참고 URL](https://kakaotech-mission2-order.pages.dev/)을 살펴보고 받는 사람 기능의 UI를 구현해요.
